requirements.txt
filterpy==1.4.5
loguru==0.7.2
PyYAML==6.0.2
torch==2.5.1+cu121
torchvision==0.20.1+cu121
h5py==3.12.1
numpy~=2.2.2
pandas==2.2.3
opencv-python~=4.10.0.84
pillow==11.0.0